requirements.txt
anyio==3.6.2
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
fastapi==0.95.2
fonttools==4.39.3
h11==0.14.0
httptools==0.5.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
lazypredict==0.2.12
lightgbm==3.3.5
matplotlib==3.7.1
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
xgboost==1.7.5
